+++ src/main/java/br/com/sistemaCadastroPersonagem/rest/controller/ParametroProficienciaController.java
@@ -33,7 +33,7 @@ public class ParametroProficienciaController {
     }
 
     @GetMapping(value = "/consulta/{id}", produces = MediaType.APPLICATION_JSON)
-    public ResponseEntity<ParametroProficienciaDTO> getById(@PathVariable Long id) {
+    public ResponseEntity<ParametroProficienciaDTO> getById(@PathVariable Integer id) {
         ParametroProficienciaDTO dto = this.service.getById(id);
         return new ResponseEntity<ParametroProficienciaDTO>(dto, HttpStatus.OK);
     }
@@ -45,14 +45,14 @@ public class ParametroProficienciaController {
     }
 
     @PutMapping(value = "/alterar/{id}", produces = MediaType.APPLICATION_JSON, consumes = MediaType.APPLICATION_JSON)
-    public ResponseEntity<ParametroProficienciaDTO> update(@PathVariable Long id, @RequestBody ParametroProficienciaDTO dtoInput) {
+    public ResponseEntity<ParametroProficienciaDTO> update(@PathVariable Integer id, @RequestBody ParametroProficienciaDTO dtoInput) {
         ParametroProficienciaDTO dto = this.service.update(id, dtoInput);
         return new ResponseEntity<ParametroProficienciaDTO>(dto, HttpStatus.NO_CONTENT);
     }
 
     @SuppressWarnings("rawtypes")
     @GetMapping(value = "/excluir/{id},{usuario}")
-    public ResponseEntity delete(@PathVariable Long id, @PathVariable String usuario) {
+    public ResponseEntity delete(@PathVariable Integer id, @PathVariable String usuario) {
         service.delete(id, usuario);
         return new ResponseEntity(HttpStatus.OK);
     }
