+++ src/main/java/br/com/sistemaCadastroPersonagem/rest/controller/RacaTracosRaciaisController.java
@@ -33,7 +33,7 @@ public class RacaTracosRaciaisController {
     }
 
     @GetMapping(value = "/consulta/{id}", produces = MediaType.APPLICATION_JSON)
-    public ResponseEntity<RacaTracosRaciaisDTO> getById(@PathVariable Long id) {
+    public ResponseEntity<RacaTracosRaciaisDTO> getById(@PathVariable Integer id) {
         RacaTracosRaciaisDTO dto = this.service.getById(id);
         return new ResponseEntity<RacaTracosRaciaisDTO>(dto, HttpStatus.OK);
     }
@@ -45,14 +45,14 @@ public class RacaTracosRaciaisController {
     }
 
     @PutMapping(value = "/alterar/{id}", produces = MediaType.APPLICATION_JSON, consumes = MediaType.APPLICATION_JSON)
-    public ResponseEntity<RacaTracosRaciaisDTO> update(@PathVariable Long id, @RequestBody RacaTracosRaciaisDTO dtoInput) {
+    public ResponseEntity<RacaTracosRaciaisDTO> update(@PathVariable Integer id, @RequestBody RacaTracosRaciaisDTO dtoInput) {
         RacaTracosRaciaisDTO dto = this.service.update(id, dtoInput);
         return new ResponseEntity<RacaTracosRaciaisDTO>(dto, HttpStatus.NO_CONTENT);
     }
 
     @SuppressWarnings("rawtypes")
     @GetMapping(value = "/excluir/{id},{usuario}")
-    public ResponseEntity delete(@PathVariable Long id, @PathVariable String usuario) {
+    public ResponseEntity delete(@PathVariable Integer id, @PathVariable String usuario) {
         service.delete(id, usuario);
         return new ResponseEntity(HttpStatus.OK);
     }
